Feature-gate the Http interface
The R2PipeHttp implementation uses reqwest, which in turn requires a lot
of dependencies. This requires building a lot of dependencies (192 with
the current reqwest version) for a feature that may not be used by
everybody.

With this commit the struct is now optional and must be enabled manually.
If the crate is built without Http support, the amount of dependencies
is now 25, greatly reducing the build time.